+++ src/Controller/PollsController.php
@@ -37,7 +37,11 @@ class PollsController extends AbstractController
      */
     public function mypolls(PollsRepository $pollsRepository): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        // deny access unless verified email
+         if (!$this->get('security.authorization_checker')->isGranted('ROLE_VERIFIED')) {
+            $this->addFlash("warning", "You must verify your email.");
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('polls/mypolls.html.twig', [
             'polls' => $pollsRepository->findAll(),
         ]);
@@ -48,7 +52,11 @@ class PollsController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        // deny access unless verified email
+        if (!$this->get('security.authorization_checker')->isGranted('ROLE_VERIFIED')) {
+            $this->addFlash("warning", "You must verify your email.");
+            return $this->redirectToRoute('app_login');
+        }
         $poll = new Polls();
         $form = $this->createForm(PollsCreateType::class, $poll);
         $form->handleRequest($request);
@@ -92,6 +100,11 @@ class PollsController extends AbstractController
      */
     public function show(Polls $poll): Response
     {
+        // deny access unless verified email
+         if (!$this->get('security.authorization_checker')->isGranted('ROLE_VERIFIED')) {
+            $this->addFlash("warning", "You must verify your email.");
+            return $this->redirectToRoute('app_login');
+        }
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         return $this->render('polls/show.html.twig', [
             'poll' => $poll,
@@ -103,6 +116,10 @@ class PollsController extends AbstractController
      */
     public function edit(Request $request, Polls $poll): Response
     {
+        if (!$this->get('security.authorization_checker')->isGranted('ROLE_VERIFIED')) {
+            $this->addFlash("warning", "You must verify your email.");
+            return $this->redirectToRoute('app_login');
+        }
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $form = $this->createForm(PollsType::class, $poll);
         $form->handleRequest($request);
@@ -124,6 +141,11 @@ class PollsController extends AbstractController
      */
     public function delete(Request $request, Polls $poll): Response
     {
+        // deny access unless admin role
+         if (!$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $this->addFlash("warning", "You must be admin.");
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$poll->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($poll);
